shift email to util
