+++ components/ContactSeller.tsx
@@ -1,12 +1,14 @@
 import ContactForm from '@/components/forms/ContactForm'
+import emailHandler from '@/utils/emailHandler'
 import { adminAuthClient } from '@/utils/supabase/admin'
-import nodemailer from 'nodemailer'
 import { z } from 'zod'
 
 export default await function ContactSeller({
   sellerID,
+  fullName,
 }: {
   sellerID: string
+  fullName: string
 }) {
   async function submit(formData: FormData) {
     'use server'
@@ -16,40 +18,22 @@ export default await function ContactSeller({
     })
     const email = formData.get('contact-email') as string
     const message = formData.get('contact-message') as string
-    const transporter = nodemailer.createTransport({
-      service: 'gmail',
-      auth: {
-        user: process.env.EMAIL,
-        pass: process.env.EMAIL_PASSWORD,
-      },
-    })
-    const { data } = await adminAuthClient.getUserById(sellerID)
-
-    const sellerEmail = data && data.user && data.user.email
-
     const { email: validatedEmail, message: validatedMessage } =
       ContactSchema.parse({ email, message })
-    const mailOptions = {
-      from: 'bamboonesttfb@gmail.com',
-      to: `${sellerEmail}`,
-      subject: 'Bamboo Nest contact',
-      text: validatedMessage,
-      html: `<p>from: ${validatedEmail}</p>
-      <p>message: ${validatedMessage}</p>`,
-      replyTo: validatedEmail,
-    }
-    transporter.sendMail(mailOptions, (error, info) => {
-      if (error) {
-        return console.log(error)
-      }
-      console.log('Email sent: ' + info.response)
-    })
+    
+    
+    // fetch seller address from auth users table
+    const { data } = await adminAuthClient.getUserById(sellerID)
+    const sellerEmail = data && data.user && data.user.email
+    if (sellerEmail) emailHandler(sellerEmail, validatedMessage, validatedEmail)
   }
 
   return (
     <>
-      <h1 className="text-center">Contact Page</h1>
+      <h1>Get in touch with {fullName}</h1>
       <ContactForm submit={submit} />
     </>
   )
 }
+
+
